+++ CHANGELOG.md
@@ -3,6 +3,9 @@ It's about time I started a changelog! This will serve from now on as the main c
 
 Note to self: See the bottom of this file for the release template text.
 
+## v1.14.5: The multipoint update, hotfix 5 (unreleased)
+- Added `//listentities`, which lists all currently loaded `ObjectRef`s. This is intended for debugging mods.
+
 
 ## v1.14.4: The multipoint update, hotfix 4 (31st July 2023)
 - When any segment of the marker wall is punched, unmark the entire wall
@@ -10,6 +13,7 @@ Note to self: See the bottom of this file for the release template text.
 
 ## v1.14.3: The multipoint update, hotfix 3 (18th July 2023)
 - Fix regions not remembering their state and being unresettable
+
 
 ## v1.14.2: The multipoint update, hotfix 2 (15th July 2023)
 - Fix crash in `//subdivide`, again due to the new position system